payment-service/payments/requirements.txt
djangorestframework
psycopg2-binary
python-dotenv
drf-yasg
requests
stripe
